+++ CoherentSolutionsTasks.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31727.386
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoherentSolutionsTasks", "CoherentSolutionsTasks\CoherentSolutionsTasks.csproj", "{1C36987B-394C-470C-BC2B-D6DE3E559C43}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoherentSolutionsTasks", "CoherentSolutionsTasks\CoherentSolutionsTasks.csproj", "{1C36987B-394C-470C-BC2B-D6DE3E559C43}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Task01-1", "Task01-1\Task01-1.csproj", "{071EC248-CE58-4A83-BDCA-C5213733F6FC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{1C36987B-394C-470C-BC2B-D6DE3E559C43}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1C36987B-394C-470C-BC2B-D6DE3E559C43}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1C36987B-394C-470C-BC2B-D6DE3E559C43}.Release|Any CPU.Build.0 = Release|Any CPU
+		{071EC248-CE58-4A83-BDCA-C5213733F6FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{071EC248-CE58-4A83-BDCA-C5213733F6FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{071EC248-CE58-4A83-BDCA-C5213733F6FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{071EC248-CE58-4A83-BDCA-C5213733F6FC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
